refactor : `MenuGroup` 패키지 변경 및 파라미터 타입 수정
