+++ src/main/java/kitchenpos/menu/application/MenuGroupService.java
@@ -1,4 +1,4 @@
-package kitchenpos.menu.ui;
+package kitchenpos.menu.application;
 
 import kitchenpos.menu.domain.MenuGroupRepository;
 import kitchenpos.menu.domain.MenuGroup;
@@ -18,10 +18,9 @@ public class MenuGroupService {
         this.menuGroupRepository = menuGroupRepository;
     }
 
-    public MenuGroupResponse create(final MenuGroupRequest menuGroupRequest) {
-        return MenuGroupResponse.of(menuGroupRepository.save(menuGroupRequest.toMenuGroup()));
+    public MenuGroupResponse create(final MenuGroup menuGroup) {
+        return MenuGroupResponse.of(menuGroupRepository.save(menuGroup));
     }
-
 
     public List<MenuGroup> list() {
         return menuGroupRepository.findAll();
